Gradle 5.6.4 -> 6.3 Kotlin 1.3.61 -> 1.3.72 Fixed some warnings of IDE Divided some classes (Commands.kt, ParkourBlock.kt)
