+++ gradle.properties
@@ -1,5 +1,3 @@
 kotlin.code.style=official
 pluginGroup=com.github.noonmaru
-pluginName=ParkourMaker
-pluginMain=com.github.noonmaru.parkourmaker.plugin.ParkourMakerPlugin
-pluginVersion=0.1
+pluginVersion=0.2
