+++ src/main/kotlin/com/github/noonmaru/parkourmaker/ParkourListener.kt
@@ -1,5 +1,6 @@
 package com.github.noonmaru.parkourmaker
 
+import com.github.noonmaru.parkourmaker.ParkourMaker.traceur
 import com.sk89q.worldedit.math.BlockVector3
 import org.bukkit.block.BlockFace
 import org.bukkit.entity.EntityType
@@ -30,11 +31,10 @@ class ParkourListener : Listener {
     @EventHandler
     fun onInteract(event: PlayerInteractEvent) {
         event.clickedBlock?.let { block ->
-            val traceur = event.player.traceur
-            traceur.challenge?.let { challenge ->
+            event.player.traceur.challenge?.let { challenge ->
                 challenge.dataByBlock[block]?.run {
                     event.isCancelled = true
-                    onInteract(challenge, traceur, event)
+                    changeState(challenge)
                 }
             }
         }
@@ -51,7 +51,7 @@ class ParkourListener : Listener {
                 traceur.challenge?.let { challenge ->
                     challenge.dataByBlock[block]?.run {
                         projectile.remove()
-                        onHit(challenge, traceur, event)
+                        changeState(challenge)
                     }
                 }
             }
@@ -61,17 +61,16 @@ class ParkourListener : Listener {
     @EventHandler
     fun onMove(event: PlayerMoveEvent) {
         event.player.let { player ->
-
             val traceur = player.traceur
             traceur.challenge?.let { challenge ->
                 val passBlock = event.to.block
                 challenge.dataByBlock[passBlock]?.run {
-                    onPass(challenge, traceur, event)
+                    onPass(traceur)
                 }
                 if (player.isOnGround) {
                     val stepBlock = passBlock.getRelative(BlockFace.DOWN)
                     challenge.dataByBlock[stepBlock]?.run {
-                        onStep(challenge, traceur, event)
+                        onStep(traceur)
                     }
                 }
             }
@@ -83,11 +82,11 @@ class ParkourListener : Listener {
         ParkourMaker.levels.values.forEach level@{ level ->
             event.blockList().forEach { block ->
                 if (level.region.contains(BlockVector3.at(block.x, block.y, block.z)))
-                    level.challenge?.let {
-                        it.dataByBlock[block]?.run {
+                    level.challenge?.let { challenge ->
+                        challenge.dataByBlock[block]?.run {
                             event.entity.remove()
                             event.isCancelled = true
-                            onExplode(it)
+                            changeState(challenge)
                             return@level
                         }
                     }
@@ -102,12 +101,12 @@ class ParkourListener : Listener {
             val fallPoint = block.getRelative(BlockFace.DOWN)
             ParkourMaker.levels.values.forEach { level ->
                 if (level.region.contains(BlockVector3.at(fallPoint.x, fallPoint.y, fallPoint.z)))
-                    level.challenge?.let {
-                        it.dataByBlock[fallPoint]?.run {
+                    level.challenge?.let { challenge ->
+                        challenge.dataByBlock[fallPoint]?.run {
                             event.entity.remove()
                             event.isCancelled = true
                             block.breakNaturally()
-                            onAnvil(it)
+                            changeState(challenge)
                         }
                     }
             }
