+++ src/main/kotlin/com/github/noonmaru/parkourmaker/block/Respawnable.kt
@@ -1,0 +1,7 @@
+package com.github.noonmaru.parkourmaker.block
+
+import org.bukkit.Location
+
+interface Respawnable {
+    val respawn: Location
+}
