+++ travis.yml
@@ -1,0 +1,19 @@
+language: java
+os: linux
+dist: xenial
+jdk:
+  - openjdk8
+before_install:
+  - chmod +x ./gradlew
+cache:
+  directories:
+    - '$HOME/.m2'
+    - '$HOME/.gradle'
+after_success:
+  - wget https://raw.githubusercontent.com/DiscordHooks/travis-ci-discord-webhook/master/send.sh
+  - chmod +x send.sh
+  - ./send.sh success $WEBHOOK_URL
+after_failure:
+  - wget https://raw.githubusercontent.com/DiscordHooks/travis-ci-discord-webhook/master/send.sh
+  - chmod +x send.sh
+  - ./send.sh failure $WEBHOOK_URL
